Character tab changes:
add the following sample archetype:
Archetype: Pinioned (15 pts)
Source: Technological
Permission: Power Theme
Intrinsics: Mutable, No Willpower No Way
This is the archetype for characters who came out of the U.S. government’s defunct, top-secret “super soldier” program, Project PINION. It aimed to create superhumans like mutants but built to specification. It didn’t always work out like that. Only a handful of subjects survived with their bodies and minds intact.
A character with this archetype has a set of superpowers that are typically related to each other or to some core theme or concept, and usually that concept is tied somehow to the character’s self-image. Sometimes that self-image is private, and sometimes it’s subconscious, not even known to the character. Since psyche and Talent are so closely linked, the powers sometimes change and evolve. But by the same token, when these characters are at their emotional worst their powers fail them entirely.